Names of all Excel functions

A dataset containing the names of all functions available in Excel. This is useful for identifying user-defined functions in formulas tokenized by xlex().

Usage

excel_functions

Format

A character vector of length 600.

Source

Pages 26--27 of Microsoft's document "Excel (.xlsx) extensions to the office openxml spreadsheetml file format p.24" https://docs.microsoft.com/en-us/openspecs/office_standards/ms-xlsx/2c5dee00-eff2-4b22-92b6-0738acd4475e, revision 8.0 2017-06-20.

Details

Note that this includes future function names that are already reserved.