Import xlsx (Excel) cell contents into a tidy structure.

tidy_xlsx() imports data from spreadsheets without coercing it into a rectangle. Each cell is represented by a row in a data frame, giving the cell's address, contents, formula, height, width, and keys to look up the cell's formatting in the return value of tidyxl::xlsx_formats().

xlsx_cells(path, sheets = NA)

Arguments

path
Path to the xlsx file.

sheets
Sheets to read. Either a character vector (the names of the sheets), an integer vector (the positions of the sheets), or NA (default, all sheets).

Value

A list of data frames. One data frame per sheet, one row per cell.

Each sheet's data is returned as a data frames, one per sheet, by the sheet name. For example, the data in a sheet named 'My Worksheet' is in x$My Worksheet. Each data frame has the following columns:

address The cell address in A1 notation.

row The row number of a cell address (integer).

col The column number of a cell address (integer).

content The content of a cell before type inference (see 'Details').

formula The formula in a cell (see 'Details').

formula_type NA for ordinary formulas, or 'array' for array formulas.

formula_ref The address (in A1 notation) of the cell that defines the formula of this cell (see 'Details').

formula_group The formula group to which the cell belongs (see 'Details').

formula_ref The address of a range of cells group to which an array formula or shared formula applies (see 'Details').

formula_group An index of a group of cells to which a shared formula applies (see 'Details').

type The type of a cell in Excel's notation (b = boolean, e = error, s = string, str = formula).

data_type The type of a cell, referring to the following columns (error, logical, numeric, date).

error The error value of a cell.

logical The boolean value of a cell.

numeric The numeric value of a cell.

date The date value of a cell.

character The string value of a cell.

comment The text of a comment attached to a cell.

height The height of a cell's row, in Excel's units.

width The width of a cell's column, in Excel's units.

style_format An index into a table of style formats tidyxl::xlsx_formats() (see 'Details').

local_format_id An index into a table of local cell formats [tidyxl::xlsx_formats()` (see 'Details').

Cell formatting is returned in tidyxl::xlsx_formats(). There are two types or scopes of formatting: 'style' formatting, such as Excel's built-in styles 'normal', 'bad', etc., and 'local' formatting, which overrides particular elements of the style, e.g. by making it bold. Both types are returned, in the $style and $local sublists of tidyxl::xlsx_formats(), with identical structures. To look up the local formatting of a given cell, take the cell's 'local_format_id' value (my_cells$data$Sheet1[1, "local_format_id"]), and use it as an index into the format structure. E.g. to look up the font size, my_formats$local$font$size[local_format_id]. To see all available formats, type str(my_formats$local).

[tidyxl::xlsx_formats()` (see 'Details').

Cell formatting is returned in tidyxl::xlsx_formats(): R:tidyxl::xlsx_formats() tidyxl::xlsx_formats(): R:tidyxl::xlsx_formats() [1, "local_format_id"]: R:1, [local_format_id]: R:local_format_id

Details

A cell has two 'values': its content, and sometimes also a formula. It also has formatting applied at the 'style' level, which can be locally overridden.

Content
Depending on the cell, the content may be a numeric value such as 365 or 365.25, it may represent a date/datetime in one of Excel's date/datetime systems, or it may be an index into an internal table of strings. tidy_xlsx() attempts to infer the correct data type of each cell, returning its value in the appropriate column (error, logical, numeric, date, character). In case this cleverness is unhelpful, the unparsed value and type information is available in the 'content' and 'type' columns.
Formula
When a cell has a formula, the value in the 'content' column is the result of the formula the last time it was evaluated. Certain groups of cells may share a formula that differs only by addresses referred to in the formula; such groups are identified by an index, the 'formula_group'. The xlsx (Excel) file format only records the formula against one cell in any group. tidy_xlsx() propogates such formulas to the other cells in a group, making the necessary changes to relative addresses in the formula. Array formulas may also apply to a group of cells, identified by an address 'formula_ref', but xlsx (Excel) file format only records the formula against one cell in the group. tidy_xlsx() propogates such formulas to the other cells in a group. Unlike shared formulas, no changes to addresses in array formulas are necessary. Formulas that refer to other workbooks currently do not name the workbooks directly, instead via indices such as [1]. It is planned to dereference these.
Formatting
Cell formatting is returned by tidyxl::xlsx_formats(). There are two types of formatting: 'style' formatting, such as Excel's built-in styles 'normal', 'bad', etc., and 'local' formatting, which overrides the style. These are returned in the $style and $local sublists of tidyxl::xlsx_formats(), with identical structures. To look up the local formatting of a given cell, take the cell's local_format_id value (my_cells$Sheet1[1, "local_format_id"]), and use it as an index into the format structure. E.g. to look up the font size, my_formats$local$font$size[local_format_id]. To see all available formats, type str(my_formats$local).
Examples

examples <- system.file("extdata/examples.xlsx", package = "tidyxl")

# All sheets
str(xlsx_cells(examples))

#> List of 3
#>  $ Sheet1                         :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	234 obs. of  20 variables:
#>   ..$ address        : chr [1:234] "A1" "B1" "A2" "B2" ...
#>   ..$ row            : int [1:234] 1 1 2 2 3 3 4 4 5 5 ...
#>   ..$ col            : int [1:234] 1 2 1 2 1 2 1 2 1 2 ...
#>   ..$ content        : chr [1:234] "#DIV/0!" "57" "1" "55" ...
#>   ..$ formula        : chr [1:234] "1/0" NA NA NA ...
#>   ..$ formula_type   : chr [1:234] NA NA NA NA ...
#>   ..$ formula_ref    : chr [1:234] NA NA NA NA ...
#>   ..$ formula_group  : int [1:234] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ type           : chr [1:234] "e" "s" "b" "s" ...
#>   ..$ data_type      : chr [1:234] "error" "character" "logical" "character" ...
#>   ..$ error          : chr [1:234] "#DIV/0!" NA NA NA ...
#>   ..$ logical        : logi [1:234] NA NA TRUE NA FALSE NA ...
#>   ..$ numeric        : num [1:234] NA NA NA NA NA ...
#>   ..$ date           : POSIXct[1:234], format: NA NA ...
#>   ..$ character      : chr [1:234] NA "Error" NA "Boolean (true)" ...
#>   ..$ comment        : chr [1:234] NA NA NA NA ...
#>   ..$ height         : num [1:234] 14.2 14.2 14.2 14.2 14.2 ...
#>   ..$ width          : num [1:234] 19.4 68 19.4 68 19.4 ...
#>   ..$ style_format   : chr [1:234] "Normal" "Normal" "Normal" "Normal" ...
#>   ..$ local_format_id: int [1:234] 1 1 1 1 1 1 1 1 5 1 ...
#>  $ 1~`!@#$%^&()_-+={}|;"'<,>.£¹÷¦°:Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	2 obs. of  20 variables:
#>   ..$ address        : chr [1:2] "B2" "B6"
#>   ..$ row            : int [1:2] 2 6
#>   ..$ col            : int [1:2] 2 2
#>   ..$ content        : chr [1:2] "30" "31"
#>   ..$ formula        : chr [1:2] NA NA
#>   ..$ formula_type   : chr [1:2] NA NA
#>   ..$ formula_ref    : chr [1:2] NA NA
#>   ..$ formula_group  : int [1:2] NA NA
#>   ..$ type           : chr [1:2] "s" "s"
#>   ..$ data_type      : chr [1:2] "character" "character"
#>   ..$ error          : chr [1:2] NA NA
#>   ..$ logical        : logi [1:2] NA NA
#>   ..$ numeric        : num [1:2] NA NA
#>   ..$ date           : POSIXct[1:2], format: NA NA
#>   ..$ character      : chr [1:2] "empty first row and col" "This worksheet has a nasty name but can be imported nevertheless"
#>   ..$ comment        : chr [1:2] "nacnudus:\nComment3" NA
#>   ..$ height         : num [1:2] 14.2 14.2
#>   ..$ width          : num [1:2] 8.38 8.38
#>   ..$ style_format   : chr [1:2] "Normal" "Normal"
#>   ..$ local_format_id: int [1:2] 1 1
#>  $ E09904.2                       :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	0 obs. of  20 variables:
#>   ..$ address        : chr(0) 
#>   ..$ row            : int(0) 
#>   ..$ col            : int(0) 
#>   ..$ content        : chr(0) 
#>   ..$ formula        : chr(0) 
#>   ..$ formula_type   : chr(0) 
#>   ..$ formula_ref    : chr(0) 
#>   ..$ formula_group  : int(0) 
#>   ..$ type           : chr(0) 
#>   ..$ data_type      : chr(0) 
#>   ..$ error          : chr(0) 
#>   ..$ logical        : logi(0) 
#>   ..$ numeric        : num(0) 
#>   ..$ date           :Classes 'POSIXct', 'POSIXt'  atomic (0) 
#>   ..- attr(*, "tzone")= chr "UTC"
#>   ..$ character      : chr(0) 
#>   ..$ comment        : chr(0) 
#>   ..$ height         : num(0) 
#>   ..$ width          : num(0) 
#>   ..$ style_format   : chr(0) 
#>   ..$ local_format_id: int(0) 

# Specific sheet either by position or by name
str(tidy_xlsx(examples, 2))

#> List of 2
#>  $ data   :List of 1
#>   ..$ 1~`!@#$%^&()_-+={}|;"'<,>.£¹÷¦°:Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	2 obs. of  20 variables:
#>   .. ..$ address        : chr [1:2] "B2" "B6"
#>   .. ..$ row            : int [1:2] 2 6
#>   .. ..$ col            : int [1:2] 2 2
#>   .. ..$ content        : chr [1:2] "30" "31"
#>   .. ..$ formula        : chr [1:2] NA NA
#>   .. ..$ formula_type   : chr [1:2] NA NA
#>   .. ..$ formula_ref    : chr [1:2] NA NA
#>   .. ..$ formula_group  : int [1:2] NA NA
#>   .. ..$ type           : chr [1:2] "s" "s"
#>   .. ..$ data_type      : chr [1:2] "character" "character"
#>   .. ..$ error          : chr [1:2] NA NA
#>   .. ..$ logical        : logi [1:2] NA NA
#>   .. ..$ numeric        : num [1:2] NA NA
#>   .. ..$ date           : POSIXct[1:2], format: NA NA
#>   .. ..$ character      : chr [1:2] "empty first row and col" "This worksheet has a nasty name but can be imported nevertheless"
#>   .. ..$ comment        : chr [1:2] "nacnudus:\nComment3" NA
#>   .. ..$ height         : num [1:2] 14.2 14.2
#>   .. ..$ width          : num [1:2] 8.38 8.38
#>   .. ..$ style_format   : chr [1:2] "Normal" "Normal"
#>   .. ..$ local_format_id: int [1:2] 1 1
#>  $ formats:List of 2
#>   ..$ local:List of 4
#>   .. ..$ numFmt: chr [1:77] "General" "d-mmm-yy" "[$-1409]d\\ mmmm\\ yyyy;@" "yyyy\\ mmmm\\ dddd" ...
#>   .. ..$ font  :List of 10
#>   .. .. ..$ bold     : logi [1:77] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   .. .. ..$ italic   : logi [1:77] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   .. .. ..$ underline: chr [1:77] NA NA NA NA ...
#>   .. .. ..$ strike   : logi [1:77] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   .. .. ..$ vertAlign: chr [1:77] NA NA NA NA ...
#>   .. .. ..$ size     : int [1:77] 11 11 11 11 11 11 11 11 11 11 ...
#>   .. .. ..$ color    :List of 4
#>   .. .. .. ..$ rgb    : chr [1:77] NA NA NA NA ...
#>   .. .. .. ..$ theme  : int [1:77] NA NA NA NA NA NA NA NA 5 NA ...
#>   .. .. .. ..$ indexed: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ tint   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ name     : chr [1:77] "Arial" "Arial" "Arial" "Arial" ...
#>   .. .. ..$ family   : int [1:77] 2 2 2 2 2 2 2 2 2 2 ...
#>   .. .. ..$ scheme   : chr [1:77] NA NA NA NA ...
#>   .. ..$ fill  :List of 2
#>   .. .. ..$ patternFill :List of 3
#>   .. .. .. ..$ fgColor    :List of 4
#>   .. .. .. .. ..$ rgb    : chr [1:77] NA NA NA NA ...
#>   .. .. .. .. ..$ theme  : int [1:77] NA NA NA NA NA NA NA 5 NA NA ...
#>   .. .. .. .. ..$ indexed: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ tint   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ bgColor    :List of 4
#>   .. .. .. .. ..$ rgb    : chr [1:77] NA NA NA NA ...
#>   .. .. .. .. ..$ theme  : int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ indexed: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ tint   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ patternType: chr [1:77] "none" "none" "none" "none" ...
#>   .. .. ..$ gradientFill:List of 8
#>   .. .. .. ..$ type   : chr [1:77] NA NA NA NA ...
#>   .. .. .. ..$ degree_: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ left   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ right  : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ top    : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ bottom : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ color1 :List of 4
#>   .. .. .. .. ..$ rgb    : chr [1:77] NA NA NA NA ...
#>   .. .. .. .. ..$ theme  : int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ indexed: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ tint   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ color2 :List of 4
#>   .. .. .. .. ..$ rgb    : chr [1:77] NA NA NA NA ...
#>   .. .. .. .. ..$ theme  : int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ indexed: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ tint   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. ..$ border:List of 12
#>   .. .. ..$ diagonalDown: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ diagonalUp  : int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ outline     : int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ left        :List of 2
#>   .. .. .. ..$ style: chr [1:77] NA NA NA NA ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : chr [1:77] NA NA NA NA ...
#>   .. .. .. .. ..$ theme  : int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ indexed: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ tint   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ right       :List of 2
#>   .. .. .. ..$ style: chr [1:77] NA NA NA NA ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : chr [1:77] NA NA NA NA ...
#>   .. .. .. .. ..$ theme  : int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ indexed: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ tint   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ start       :List of 2
#>   .. .. .. ..$ style: chr [1:77] NA NA NA NA ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : chr [1:77] NA NA NA NA ...
#>   .. .. .. .. ..$ theme  : int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ indexed: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ tint   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ end         :List of 2
#>   .. .. .. ..$ style: chr [1:77] NA NA NA NA ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : chr [1:77] NA NA NA NA ...
#>   .. .. .. .. ..$ theme  : int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ indexed: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ tint   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ top         :List of 2
#>   .. .. .. ..$ style: chr [1:77] NA NA NA NA ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : chr [1:77] NA NA NA NA ...
#>   .. .. .. .. ..$ theme  : int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ indexed: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ tint   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ bottom      :List of 2
#>   .. .. .. ..$ style: chr [1:77] NA NA NA NA ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : chr [1:77] NA NA NA NA ...
#>   .. .. .. .. ..$ theme  : int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ indexed: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ tint   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ diagonal    :List of 2
#>   .. .. .. ..$ style: chr [1:77] NA NA NA NA ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : chr [1:77] NA NA NA NA ...
#>   .. .. .. .. ..$ theme  : int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ indexed: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ tint   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ vertical    :List of 2
#>   .. .. .. ..$ style: chr [1:77] NA NA NA NA ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : chr [1:77] NA NA NA NA ...
#>   .. .. .. .. ..$ theme  : int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ indexed: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ tint   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ horizontal  :List of 2
#>   .. .. .. ..$ style: chr [1:77] NA NA NA NA ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : chr [1:77] NA NA NA NA ...
#>   .. .. .. .. ..$ theme  : int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ indexed: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ tint   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ style:List of 4
#>   .. ..$ numFmt: Named chr [1:6] "General" "General" "yyyy\\-mm\\-dd\\ hh:mm:ss" "0%" ...
#>   .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ font  :List of 10
#>   .. .. ..$ bold     : Named logi [1:6] FALSE FALSE FALSE FALSE FALSE FALSE
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ italic   : Named logi [1:6] FALSE FALSE FALSE FALSE FALSE FALSE
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ underline: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ strike   : Named logi [1:6] FALSE FALSE FALSE FALSE FALSE FALSE
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ vertAlign: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ size     : Named int [1:6] 11 11 11 11 11 11
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ color    :List of 4
#>   .. .. .. ..$ rgb    : Named chr [1:6] NA "FF9C0006" "FF000000" NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ theme  : Named int [1:6] NA NA 2 NA 11 2
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ name     : Named chr [1:6] "Arial" "Calibri" "Calibri" "Arial" ...
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ family   : Named int [1:6] 2 2 2 2 2 2
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ scheme   : Named chr [1:6] NA "minor" "minor" NA ...
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ fill  :List of 2
#>   .. .. ..$ patternFill :List of 3
#>   .. .. .. ..$ fgColor    :List of 4
#>   .. .. .. .. ..$ rgb    : Named chr [1:6] NA "FFFFC7CE" NA NA ...
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ bgColor    :List of 4
#>   .. .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ patternType: Named chr [1:6] "none" "solid" "none" "none" ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ gradientFill:List of 8
#>   .. .. .. ..$ type   : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ degree_: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ left   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ right  : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ top    : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ bottom : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ color1 :List of 4
#>   .. .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ color2 :List of 4
#>   .. .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ border:List of 12
#>   .. .. ..$ diagonalDown: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ diagonalUp  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ outline     : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ left        :List of 2
#>   .. .. .. ..$ style: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ right       :List of 2
#>   .. .. .. ..$ style: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ start       :List of 2
#>   .. .. .. ..$ style: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ end         :List of 2
#>   .. .. .. ..$ style: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ top         :List of 2
#>   .. .. .. ..$ style: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ bottom      :List of 2
#>   .. .. .. ..$ style: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ diagonal    :List of 2
#>   .. .. .. ..$ style: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ vertical    :List of 2
#>   .. .. .. ..$ style: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ horizontal  :List of 2
#>   .. .. .. ..$ style: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...

str(tidy_xlsx(examples, "Sheet1"))

#> List of 2
#>  $ data   :List of 1
#>   ..$ Sheet1:Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	234 obs. of  20 variables:
#>   .. ..$ address        : chr [1:234] "A1" "B1" "A2" "B2" ...
#>   .. ..$ row            : int [1:234] 1 1 2 2 3 3 4 4 5 5 ...
#>   .. ..$ col            : int [1:234] 1 2 1 2 1 2 1 2 1 2 ...
#>   .. ..$ content        : chr [1:234] "#DIV/0!" "57" "1" "55" ...
#>   .. ..$ formula        : chr [1:234] "1/0" NA NA NA ...
#>   .. ..$ formula_type   : chr [1:234] NA NA NA NA ...
#>   .. ..$ formula_ref    : chr [1:234] NA NA NA NA ...
#>   .. ..$ formula_group  : int [1:234] NA NA NA NA NA NA NA NA NA NA ...
#>   .. ..$ type           : chr [1:234] "e" "s" "b" "s" ...
#>   .. ..$ data_type      : chr [1:234] "error" "character" "logical" "character" ...
#>   .. ..$ error          : chr [1:234] "#DIV/0!" NA NA NA ...
#>   .. ..$ logical        : logi [1:234] NA NA TRUE NA FALSE NA ...
#>   .. ..$ numeric        : num [1:234] NA NA NA NA NA ...
#>   .. ..$ date           : POSIXct[1:234], format: NA NA ...
#>   .. ..$ character      : chr [1:234] NA "Error" NA "Boolean (true)" ...
#>   .. ..$ comment        : chr [1:234] NA NA NA NA ...
#>   .. ..$ height         : num [1:234] 14.2 14.2 14.2 14.2 14.2 ...
#>   .. ..$ width          : num [1:234] 19.4 68 19.4 68 19.4 ...
#>   .. ..$ style_format   : chr [1:234] "Normal" "Normal" "Normal" "Normal" ...
#>   .. ..$ local_format_id: int [1:234] 1 1 1 1 1 1 1 1 5 1 ...
#>  $ formats:List of 2
#>   ..$ local:List of 4
#>   .. ..$ numFmt: chr [1:77] "General" "d-mmm-yy" "[$-1409]d\\ mmmm\\ yyyy;@" "yyyy\\ mmmm\\ dddd" ...
#>   .. ..$ font  :List of 10
#>   .. .. ..$ bold     : logi [1:77] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   .. .. ..$ italic   : logi [1:77] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   .. .. ..$ underline: chr [1:77] NA NA NA NA ...
#>   .. .. ..$ strike   : logi [1:77] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   .. .. ..$ vertAlign: chr [1:77] NA NA NA NA ...
#>   .. .. ..$ size     : int [1:77] 11 11 11 11 11 11 11 11 11 11 ...
#>   .. .. ..$ color    :List of 4
#>   .. .. .. ..$ rgb    : chr [1:77] NA NA NA NA ...
#>   .. .. .. ..$ theme  : int [1:77] NA NA NA NA NA NA NA NA 5 NA ...
#>   .. .. .. ..$ indexed: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ tint   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ name     : chr [1:77] "Arial" "Arial" "Arial" "Arial" ...
#>   .. .. ..$ family   : int [1:77] 2 2 2 2 2 2 2 2 2 2 ...
#>   .. .. ..$ scheme   : chr [1:77] NA NA NA NA ...
#>   .. ..$ fill  :List of 2
#>   .. .. ..$ patternFill :List of 3
#>   .. .. .. ..$ fgColor    :List of 4
#>   .. .. .. .. ..$ rgb    : chr [1:77] NA NA NA NA ...
#>   .. .. .. .. ..$ theme  : int [1:77] NA NA NA NA NA NA NA 5 NA NA ...
#>   .. .. .. .. ..$ indexed: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ tint   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ bgColor    :List of 4
#>   .. .. .. .. ..$ rgb    : chr [1:77] NA NA NA NA ...
#>   .. .. .. .. ..$ theme  : int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ indexed: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ tint   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ patternType: chr [1:77] "none" "none" "none" "none" ...
#>   .. .. ..$ gradientFill:List of 8
#>   .. .. .. ..$ type   : chr [1:77] NA NA NA NA ...
#>   .. .. .. ..$ degree_: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ left   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ right  : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ top    : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ bottom : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ color1 :List of 4
#>   .. .. .. .. ..$ rgb    : chr [1:77] NA NA NA NA ...
#>   .. .. .. .. ..$ theme  : int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ indexed: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ tint   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ color2 :List of 4
#>   .. .. .. .. ..$ rgb    : chr [1:77] NA NA NA NA ...
#>   .. .. .. .. ..$ theme  : int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ indexed: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ tint   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. ..$ border:List of 12
#>   .. .. ..$ diagonalDown: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ diagonalUp  : int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ outline     : int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ left        :List of 2
#>   .. .. .. ..$ style: chr [1:77] NA NA NA NA ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : chr [1:77] NA NA NA NA ...
#>   .. .. .. .. ..$ theme  : int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ indexed: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ tint   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ right       :List of 2
#>   .. .. .. ..$ style: chr [1:77] NA NA NA NA ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : chr [1:77] NA NA NA NA ...
#>   .. .. .. .. ..$ theme  : int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ indexed: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ tint   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ start       :List of 2
#>   .. .. .. ..$ style: chr [1:77] NA NA NA NA ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : chr [1:77] NA NA NA NA ...
#>   .. .. .. .. ..$ theme  : int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ indexed: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ tint   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ end         :List of 2
#>   .. .. .. ..$ style: chr [1:77] NA NA NA NA ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : chr [1:77] NA NA NA NA ...
#>   .. .. .. .. ..$ theme  : int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ indexed: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ tint   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ top         :List of 2
#>   .. .. .. ..$ style: chr [1:77] NA NA NA NA ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : chr [1:77] NA NA NA NA ...
#>   .. .. .. .. ..$ theme  : int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ indexed: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ tint   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ bottom      :List of 2
#>   .. .. .. ..$ style: chr [1:77] NA NA NA NA ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : chr [1:77] NA NA NA NA ...
#>   .. .. .. .. ..$ theme  : int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ indexed: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ tint   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ diagonal    :List of 2
#>   .. .. .. ..$ style: chr [1:77] NA NA NA NA ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : chr [1:77] NA NA NA NA ...
#>   .. .. .. .. ..$ theme  : int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ indexed: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ tint   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ vertical    :List of 2
#>   .. .. .. ..$ style: chr [1:77] NA NA NA NA ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : chr [1:77] NA NA NA NA ...
#>   .. .. .. .. ..$ theme  : int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ indexed: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ tint   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ horizontal  :List of 2
#>   .. .. .. ..$ style: chr [1:77] NA NA NA NA ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : chr [1:77] NA NA NA NA ...
#>   .. .. .. .. ..$ theme  : int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ indexed: int [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. .. ..$ tint   : num [1:77] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ style:List of 4
#>   .. ..$ numFmt: Named chr [1:6] "General" "General" "yyyy\\-mm\\-dd\\ hh:mm:ss" "0%" ...
#>   .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ font  :List of 10
#>   .. .. ..$ bold     : Named logi [1:6] FALSE FALSE FALSE FALSE FALSE FALSE
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ italic   : Named logi [1:6] FALSE FALSE FALSE FALSE FALSE FALSE
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ underline: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ strike   : Named logi [1:6] FALSE FALSE FALSE FALSE FALSE FALSE
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ vertAlign: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ size     : Named int [1:6] 11 11 11 11 11 11
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ color    :List of 4
#>   .. .. .. ..$ rgb    : Named chr [1:6] NA "FF9C0006" "FF000000" NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ theme  : Named int [1:6] NA NA 2 NA 11 2
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ name     : Named chr [1:6] "Arial" "Calibri" "Calibri" "Arial" ...
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ family   : Named int [1:6] 2 2 2 2 2 2
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ scheme   : Named chr [1:6] NA "minor" "minor" NA ...
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ fill  :List of 2
#>   .. .. ..$ patternFill :List of 3
#>   .. .. .. ..$ fgColor    :List of 4
#>   .. .. .. .. ..$ rgb    : Named chr [1:6] NA "FFFFC7CE" NA NA ...
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ bgColor    :List of 4
#>   .. .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ patternType: Named chr [1:6] "none" "solid" "none" "none" ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ gradientFill:List of 8
#>   .. .. .. ..$ type   : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ degree_: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ left   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ right  : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ top    : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ bottom : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ color1 :List of 4
#>   .. .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ color2 :List of 4
#>   .. .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ border:List of 12
#>   .. .. ..$ diagonalDown: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ diagonalUp  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ outline     : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ left        :List of 2
#>   .. .. .. ..$ style: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ right       :List of 2
#>   .. .. .. ..$ style: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ start       :List of 2
#>   .. .. .. ..$ style: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ end         :List of 2
#>   .. .. .. ..$ style: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ top         :List of 2
#>   .. .. .. ..$ style: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ bottom      :List of 2
#>   .. .. .. ..$ style: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ diagonal    :List of 2
#>   .. .. .. ..$ style: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ vertical    :List of 2
#>   .. .. .. ..$ style: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ horizontal  :List of 2
#>   .. .. .. ..$ style: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ color:List of 4
#>   .. .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...

# The formats of particular cells can be retrieved like this:

Sheet1 <- xlsx_cells(examples)$Sheet1
formats <- xlsx_formats(examples)

my_formats$local$font$bold[Sheet1$local_format_id]

#> Error in eval(expr, envir, enclos): object 'my_formats' not found

my_formats$style$font$bold[Sheet1$style_format]

#> Error in eval(expr, envir, enclos): object 'my_formats' not found

# To filter for cells of a particular format, first filter the formats to get
# the relevant indices, and then filter the cells by those indices.
bold_indices <- which(my_formats$local$font$bold)

#> Error in which(my_formats$local$font$bold): object 'my_formats' not found

Sheet1[Sheet1$local_format_id %in% bold_indices, ]

#> Error in Sheet1$local_format_id %in% bold_indices: object 'bold_indices' not found