Import xlsx (Excel) formatting definitions.

xlsx_formats() imports formatting definitions from spreadsheets. The structure is a nested list, e.g. bold is a vector within the list font, which is within the list local, which is within the list returned by xlsx_formats(). You can look up a cell's formatting by indexing the bottom-level vectors. See 'Details' for examples.

xlsx_formats(path)

Arguments

path
Path to the xlsx file.

Value

A nested list of vectors, beginning at the top level with $style and $local, then drilling down to the vectors that hold the definitions. E.g. my_formats$local$font$size.

Details

There are two types of formatting: 'style' formatting, such as Excel's built-in styles 'normal', 'bad', etc., and 'local' formatting, which overrides the style. These are returned in the $style and $local sublists of xlsx_formats(), with identical structures.

To look up the local formatting of a given cell, take the cell's local_format_id value (my_cells$Sheet1[1, "local_format_id"]), and use it as an index into the format structure. E.g. to look up the font size, my_formats$local$font$size[local_format_id]. To see all available formats, type str(my_formats$local).

Colours may be recorded in any of three ways: a hexadecimal RGB string with or without alpha, an 'indexed' colour, and an index into a 'theme'. xlsx_formats() dereferences 'indexed' and 'theme' colours to their hexadecimal RGB string representation, and standardises all RGB strings to have an alpha channel in the first two characters. The 'index' and the 'theme' name are still provided. To filter by an RGB string, you could look up the RGB values in a spreadsheet program (e.g. Excel, LibreOffice, Gnumeric), and use the grDevices::rgb() function to convert these to a hexadecimal string.

Examples

examples <- system.file("extdata/examples.xlsx", package = "tidyxl")
str(xlsx_formats(examples))

#> List of 2
#>  $ local:List of 6
#>   ..$ numFmt    : chr [1:79] "General" "d-mmm-yy" "[$-1409]d\\ mmmm\\ yyyy;@" "yyyy\\ mmmm\\ dddd" ...
#>   ..$ font      :List of 10
#>   .. ..$ bold     : logi [1:79] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   .. ..$ italic   : logi [1:79] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   .. ..$ underline: chr [1:79] "" "" "" "" ...
#>   .. ..$ strike   : logi [1:79] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   .. ..$ vertAlign: chr [1:79] "" "" "" "" ...
#>   .. ..$ size     : num [1:79] 11 11 11 11 11 11 11 11 11 11 ...
#>   .. ..$ color    :List of 4
#>   .. .. ..$ rgb    : chr [1:79] NA NA NA NA ...
#>   .. .. ..$ theme  : int [1:79] NA NA NA NA NA NA NA NA 5 NA ...
#>   .. .. ..$ indexed: int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ tint   : num [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. ..$ name     : chr [1:79] "Arial" "Arial" "Arial" "Arial" ...
#>   .. ..$ family   : int [1:79] 2 2 2 2 2 2 2 2 2 2 ...
#>   .. ..$ scheme   : chr [1:79] NA NA NA NA ...
#>   ..$ fill      :List of 2
#>   .. ..$ patternFill :List of 3
#>   .. .. ..$ fgColor    :List of 4
#>   .. .. .. ..$ rgb    : chr [1:79] NA NA NA NA ...
#>   .. .. .. ..$ theme  : int [1:79] NA NA NA NA NA NA NA 5 NA NA ...
#>   .. .. .. ..$ indexed: int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ tint   : num [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ bgColor    :List of 4
#>   .. .. .. ..$ rgb    : chr [1:79] NA NA NA NA ...
#>   .. .. .. ..$ theme  : int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ indexed: int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ tint   : num [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ patternType: chr [1:79] "none" "none" "none" "none" ...
#>   .. ..$ gradientFill:List of 8
#>   .. .. ..$ type   : chr [1:79] NA NA NA NA ...
#>   .. .. ..$ degree_: chr [1:79] NA NA NA NA ...
#>   .. .. ..$ left   : chr [1:79] NA NA NA NA ...
#>   .. .. ..$ right  : chr [1:79] NA NA NA NA ...
#>   .. .. ..$ top    : chr [1:79] NA NA NA NA ...
#>   .. .. ..$ bottom : chr [1:79] NA NA NA NA ...
#>   .. .. ..$ color1 :List of 4
#>   .. .. .. ..$ rgb    : chr [1:79] NA NA NA NA ...
#>   .. .. .. ..$ theme  : int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ indexed: int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ tint   : num [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ color2 :List of 4
#>   .. .. .. ..$ rgb    : chr [1:79] NA NA NA NA ...
#>   .. .. .. ..$ theme  : int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ indexed: int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ tint   : num [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ border    :List of 12
#>   .. ..$ diagonalDown: int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. ..$ diagonalUp  : int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. ..$ outline     : int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. ..$ left        :List of 2
#>   .. .. ..$ style: chr [1:79] NA NA NA NA ...
#>   .. .. ..$ color:List of 4
#>   .. .. .. ..$ rgb    : chr [1:79] NA NA NA NA ...
#>   .. .. .. ..$ theme  : int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ indexed: int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ tint   : num [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. ..$ right       :List of 2
#>   .. .. ..$ style: chr [1:79] NA NA NA NA ...
#>   .. .. ..$ color:List of 4
#>   .. .. .. ..$ rgb    : chr [1:79] NA NA NA NA ...
#>   .. .. .. ..$ theme  : int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ indexed: int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ tint   : num [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. ..$ start       :List of 2
#>   .. .. ..$ style: chr [1:79] NA NA NA NA ...
#>   .. .. ..$ color:List of 4
#>   .. .. .. ..$ rgb    : chr [1:79] NA NA NA NA ...
#>   .. .. .. ..$ theme  : int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ indexed: int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ tint   : num [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. ..$ end         :List of 2
#>   .. .. ..$ style: chr [1:79] NA NA NA NA ...
#>   .. .. ..$ color:List of 4
#>   .. .. .. ..$ rgb    : chr [1:79] NA NA NA NA ...
#>   .. .. .. ..$ theme  : int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ indexed: int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ tint   : num [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. ..$ top         :List of 2
#>   .. .. ..$ style: chr [1:79] NA NA NA NA ...
#>   .. .. ..$ color:List of 4
#>   .. .. .. ..$ rgb    : chr [1:79] NA NA NA NA ...
#>   .. .. .. ..$ theme  : int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ indexed: int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ tint   : num [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. ..$ bottom      :List of 2
#>   .. .. ..$ style: chr [1:79] NA NA NA NA ...
#>   .. .. ..$ color:List of 4
#>   .. .. .. ..$ rgb    : chr [1:79] NA NA NA NA ...
#>   .. .. .. ..$ theme  : int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ indexed: int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ tint   : num [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. ..$ diagonal    :List of 2
#>   .. .. ..$ style: chr [1:79] NA NA NA NA ...
#>   .. .. ..$ color:List of 4
#>   .. .. .. ..$ rgb    : chr [1:79] NA NA NA NA ...
#>   .. .. .. ..$ theme  : int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ indexed: int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ tint   : num [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. ..$ vertical    :List of 2
#>   .. .. ..$ style: chr [1:79] NA NA NA NA ...
#>   .. .. ..$ color:List of 4
#>   .. .. .. ..$ rgb    : chr [1:79] NA NA NA NA ...
#>   .. .. .. ..$ theme  : int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ indexed: int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ tint   : num [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. ..$ horizontal  :List of 2
#>   .. .. ..$ style: chr [1:79] NA NA NA NA ...
#>   .. .. ..$ color:List of 4
#>   .. .. .. ..$ rgb    : chr [1:79] NA NA NA NA ...
#>   .. .. .. ..$ theme  : int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ indexed: int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. .. ..$ tint   : num [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ alignment :List of 8
#>   .. ..$ horizontal     : chr [1:79] NA NA NA NA ...
#>   .. ..$ vertical       : chr [1:79] NA NA NA NA ...
#>   .. ..$ wrapText       : logi [1:79] NA NA NA NA NA NA ...
#>   .. ..$ readingOrder   : chr [1:79] NA NA NA NA ...
#>   .. ..$ indent         : int [1:79] NA NA NA NA NA NA NA NA NA 1 ...
#>   .. ..$ justifyLastLine: logi [1:79] NA NA NA NA NA NA ...
#>   .. ..$ shrinkToFit    : logi [1:79] NA NA NA NA NA NA ...
#>   .. ..$ textRotation   : int [1:79] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ protection:List of 2
#>   .. ..$ locked: logi [1:79] NA NA NA NA NA NA ...
#>   .. ..$ hidden: logi [1:79] NA NA NA NA NA NA ...
#>  $ style:List of 6
#>   ..$ numFmt    : Named chr [1:6] "General" "General" "yyyy\\-mm\\-dd\\ hh:mm:ss" "0%" ...
#>   .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   ..$ font      :List of 10
#>   .. ..$ bold     : Named logi [1:6] FALSE FALSE FALSE FALSE FALSE FALSE
#>   .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ italic   : Named logi [1:6] FALSE FALSE FALSE FALSE FALSE FALSE
#>   .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ underline: Named chr [1:6] "" "" "" "" ...
#>   .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ strike   : Named logi [1:6] FALSE FALSE FALSE FALSE FALSE FALSE
#>   .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ vertAlign: Named chr [1:6] "" "" "" "" ...
#>   .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ size     : Named num [1:6] 11 11 11 11 11 11
#>   .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ color    :List of 4
#>   .. .. ..$ rgb    : Named chr [1:6] NA "FF9C0006" "FF000000" NA ...
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ theme  : Named int [1:6] NA NA 2 NA 11 2
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ name     : Named chr [1:6] "Arial" "Calibri" "Calibri" "Arial" ...
#>   .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ family   : Named int [1:6] 2 2 2 2 2 2
#>   .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ scheme   : Named chr [1:6] NA "minor" "minor" NA ...
#>   .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   ..$ fill      :List of 2
#>   .. ..$ patternFill :List of 3
#>   .. .. ..$ fgColor    :List of 4
#>   .. .. .. ..$ rgb    : Named chr [1:6] NA "FFFFC7CE" NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ bgColor    :List of 4
#>   .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ patternType: Named chr [1:6] "none" "solid" "none" "none" ...
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ gradientFill:List of 8
#>   .. .. ..$ type   : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ degree_: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ left   : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ right  : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ top    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ bottom : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ color1 :List of 4
#>   .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ color2 :List of 4
#>   .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   ..$ border    :List of 12
#>   .. ..$ diagonalDown: Named int [1:6] NA NA NA NA NA NA
#>   .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ diagonalUp  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ outline     : Named int [1:6] NA NA NA NA NA NA
#>   .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ left        :List of 2
#>   .. .. ..$ style: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ color:List of 4
#>   .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ right       :List of 2
#>   .. .. ..$ style: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ color:List of 4
#>   .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ start       :List of 2
#>   .. .. ..$ style: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ color:List of 4
#>   .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ end         :List of 2
#>   .. .. ..$ style: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ color:List of 4
#>   .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ top         :List of 2
#>   .. .. ..$ style: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ color:List of 4
#>   .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ bottom      :List of 2
#>   .. .. ..$ style: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ color:List of 4
#>   .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ diagonal    :List of 2
#>   .. .. ..$ style: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ color:List of 4
#>   .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ vertical    :List of 2
#>   .. .. ..$ style: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ color:List of 4
#>   .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ horizontal  :List of 2
#>   .. .. ..$ style: Named chr [1:6] NA NA NA NA ...
#>   .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. ..$ color:List of 4
#>   .. .. .. ..$ rgb    : Named chr [1:6] NA NA NA NA ...
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ theme  : Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ indexed: Named int [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. .. .. ..$ tint   : Named num [1:6] NA NA NA NA NA NA
#>   .. .. .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   ..$ alignment :List of 8
#>   .. ..$ horizontal     : Named chr [1:6] NA NA "center" NA ...
#>   .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ vertical       : Named chr [1:6] NA NA NA NA ...
#>   .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ wrapText       : Named logi [1:6] NA NA TRUE NA NA NA
#>   .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ readingOrder   : Named chr [1:6] NA NA "left-to-right" NA ...
#>   .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ indent         : Named int [1:6] NA NA NA NA NA NA
#>   .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ justifyLastLine: Named logi [1:6] NA NA TRUE NA NA NA
#>   .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ shrinkToFit    : Named logi [1:6] NA NA TRUE NA NA NA
#>   .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ textRotation   : Named int [1:6] NA NA NA NA NA NA
#>   .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   ..$ protection:List of 2
#>   .. ..$ locked: Named logi [1:6] NA NA FALSE NA NA NA
#>   .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...
#>   .. ..$ hidden: Named logi [1:6] NA NA TRUE NA NA NA
#>   .. .. ..- attr(*, "names")= chr [1:6] "Normal" "Bad" "styledate" "Percent" ...

# The formats of particular cells can be retrieved like this:

cells <- xlsx_cells(examples)
formats <- xlsx_formats(examples)

formats$local$font$bold[cells$local_format_id]

#>   [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [13] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [25] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [37] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [49] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE
#>  [61] FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [73] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [85] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE
#>  [97] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [109] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [121] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [133] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [145] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [157] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [169] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [181] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [193] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [205] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [217] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [229] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [241] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE

formats$style$font$bold[cells$style_format]

#>    Normal    Normal    Normal    Normal    Normal    Normal    Normal    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>    Normal    Normal    Normal    Normal    Normal    Normal    Normal    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>    Normal    Normal    Normal    Normal    Normal    Normal   Percent    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>    Normal    Normal    Normal    Normal    Normal    Normal    Normal    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>    Normal    Normal    Normal    Normal    Normal    Normal    Normal    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>    Normal    Normal    Normal    Normal    Normal    Normal    Normal    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>    Normal    Normal    Normal    Normal    Normal    Normal    Normal    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>    Normal    Normal    Normal    Normal       Bad    Normal       Bad    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#> styledate    Normal    Normal    Normal    Normal    Normal    Normal    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>    Normal    Normal styledate    Normal    Normal    Normal    Normal    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>    Normal    Normal    Normal    Normal    Normal    Normal    Normal    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>    Normal    Normal    Normal    Normal    Normal    Normal    Normal    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>    Normal    Normal    Normal    Normal    Normal  Normal 2    Normal  Normal 2 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>  Normal 2  Normal 2  Normal 2  Normal 2  Normal 2  Normal 2  Normal 2    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>  Normal 2    Normal  Normal 2    Normal  Normal 2    Normal  Normal 2    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>  Normal 2    Normal  Normal 2    Normal  Normal 2    Normal  Normal 2    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>  Normal 2 Hyperlink  Normal 2 Hyperlink  Normal 2    Normal  Normal 2    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>    Normal    Normal    Normal    Normal    Normal    Normal    Normal    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>    Normal    Normal    Normal    Normal    Normal    Normal    Normal    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>    Normal    Normal    Normal    Normal    Normal    Normal    Normal    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>    Normal    Normal    Normal    Normal    Normal    Normal    Normal    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>    Normal    Normal    Normal    Normal    Normal    Normal    Normal    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>    Normal    Normal    Normal    Normal    Normal    Normal    Normal    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>    Normal    Normal    Normal    Normal    Normal    Normal    Normal    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>    Normal    Normal    Normal    Normal    Normal    Normal    Normal    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>    Normal    Normal    Normal    Normal    Normal    Normal    Normal    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>    Normal    Normal    Normal    Normal    Normal    Normal    Normal    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>    Normal    Normal    Normal    Normal    Normal    Normal    Normal    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>    Normal    Normal    Normal    Normal    Normal    Normal    Normal    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>    Normal    Normal    Normal    Normal    Normal    Normal    Normal    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>    Normal    Normal    Normal    Normal    Normal    Normal    Normal    Normal 
#>     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE     FALSE 
#>    Normal 
#>     FALSE 

# To filter for cells of a particular format, first filter the formats to get
# the relevant indices, and then filter the cells by those indices.
bold_indices <- which(formats$local$font$bold)
cells[cells$local_format_id %in% bold_indices, ]

#> # A tibble: 3 x 21
#>    sheet address   row   col is_blank data_type error logical numeric   date
#>    <chr>   <chr> <int> <int>    <lgl>     <chr> <chr>   <lgl>   <dbl> <dttm>
#> 1 Sheet1     A28    28     1    FALSE character  <NA>      NA      NA     NA
#> 2 Sheet1     A32    32     1    FALSE character  <NA>      NA      NA     NA
#> 3 Sheet1     A49    49     1    FALSE character  <NA>      NA      NA     NA
#> # ... with 11 more variables: character <chr>, character_formatted <list>,
#> #   formula <chr>, is_array <lgl>, formula_ref <chr>, formula_group <int>,
#> #   comment <chr>, height <dbl>, width <dbl>, style_format <chr>,
#> #   local_format_id <int>