Import xlsx (Excel) cell contents into a tidy structure.

tidy_xlsx imports data from spreadsheets without coercing it into a rectangle. Each cell is represented by a row in a data frame, giving the cell's address, contents, formula, height, width, and keys to look up the cell's formatting in an adjacent data structure within the list returned by this function.

tidy_xlsx(path, sheets = NA)

Arguments

path

Path to the xlsx file.

sheets

Sheets to read. Either a character vector (the names of the sheets), an integer vector (the positions of the sheets), or NA (default, all sheets).

Value

A list of the data within each sheet ($data), and the formatting applied to each cell ($formats).

Each sheet's data is returned as a data frames, one per sheet, by the sheet name. For example, the data in a sheet named 'My Worksheet' is in x$data$`My Worksheet`. Each data frame has the following columns:

address

The cell address in A1 notation.

row

The row number of a cell address (integer).

col

The column numer of a cell address (integer).

content

The content of a cell before type inference (see 'Details').

formula

The formula in a cell (see 'Details').

formula_type

NA for ordinary formulas, or 'array' for array formulas.

formula_ref

The address (in A1 notation) of the cell that defines the formula of this cell (see 'Details').

formula_group

The formula group to which the cell belongs (see 'Details').

formula_ref

The address of a range of cells group to which an array formula or shared formula applies (see 'Details').

formula_group

An index of a group of cells to which a shared formula applies (see 'Details').

type

The type of a cell in Excel's notation (b = boolean, e = error, s = string, str = formula).

data_type

The type of a cell, referring to the following columns (error, logical, numeric, date).

error

The error value of a cell.

logical

The boolean value of a cell.

numeric

The numeric value of a cell.

date

The date value of a cell.

character

The string value of a cell.

comment

The text of a comment attached to a cell.

height

The height of a cell's row, in Excel's units.

width

The width of a cell's column, in Excel's units.

style_format

An index into a table of style formats x$formats$style (see 'Details').

local_format_id

An index into a table of local cell formats x$formats$local (see 'Details').

Cell formatting is returned in x$formats. There are two types or scopes of formatting: 'style' formatting, such as Excel's built-in styles 'normal', 'bad', etc., and 'local' formatting, which overrides particular elements of the style, e.g. by making it bold. Both types of are returned in x$formats$style and x$formats$local, with identical structures. To look up the local formatting of a given cell, take the cell's 'local_format_id' value (x$data$Sheet1[1, "local_format_id"]), and use it as an index into the format structure. E.g. to look up the font size, x$formats$local$font$size[local_format_id]. To see all available formats, type `str(x$formats$local)`.

Colours may be recorded in any of three ways: a hexadecimal RGB string with or without alpha, an 'indexed' colour, and an index into a 'theme'. tidy_xlsx dereferences 'indexed' and 'theme' colours to their hexadecimal RGB string representation, and standardises all RGB strings to have an alpha channel in the first two characters. The 'index' and the 'theme' name are still provided. To filter by an RGB string, you could look up the RGB values in a spreadsheet program (e.g. Excel, LibreOffice, Gnumeric), and use the rgb function to convert these to a hexadecimal string.

Details

A cell has two 'values': its content, and sometimes also a formula. It also has formatting applied at the 'style' level, which can be locally overriden.

content

Depending on the cell, the content may be a numeric value such as 365 or 365.25, it may represent a date/datetime in one of Excel's date/datetime systems, or it may be an index into an internal table of strings. tidy_xlsx attempts to infer the correct data type of each cell, returning its value in the appropriate column (error, logical, numeric, date, character). In case this cleverness is unhelpful, the unparsed value and type information is available in the 'content' and 'type' columns.

formula

When a cell has a formula, the value in the 'content' column is the result of the formula the last time it was evaluated.

Certain groups of cells may share a formula that differs only by addresses refered to in the formula; such groups are identified by an index, the 'formula_group'. The xlsx (Excel) file format only records the formula against one cell in any group. It is planned for tidy_xlsx to parse such formulas and copy them to the other cells in a group, making the necessary changes to addresses in the formula.

Array formulas may also apply to a group of cells, identified by an address 'formula_ref', but xlsx (Excel) file format only records the formula against one cell in the group. It is planned for tidy_xlsx to parse such addresses and copy the array formula to the other cells in the group. Unlike shared formulas, no changes to addresses in array formulas are necessary.

Formulas that refer to other workbooks currently do not name the workbooks directly, instead via indices such as [1]. It is planned to dereference these.

formatting

Cell formatting is returned in x$formats. There are two types of formatting: 'style' formatting, such as Excel's built-in styles 'normal', 'bad', etc., and 'local' formatting, which overrides the style. These are returned in x$formats$style and x$formats$local, with identical structures. To look up the local formatting of a given cell, take the cell's 'local_format_id' value (x$data$Sheet1[1, "local_format_id"]), and use it as an index into the format structure. E.g. to look up the font size, x$formats$local$font$size[local_format_id]. To see all available formats, type `str(x$formats$local)`.

Examples

examples <- system.file("extdata/examples.xlsx", package = "tidyxl")

# All sheets
str(tidy_xlsx(examples)$data)

#> List of 2
#>  $ Sheet1                         :Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	183 obs. of  20 variables:
#>   ..$ address        : chr [1:183] "A1" "B1" "A2" "B2" ...
#>   ..$ row            : int [1:183] 1 1 2 2 3 3 4 4 5 5 ...
#>   ..$ col            : int [1:183] 1 2 1 2 1 2 1 2 1 2 ...
#>   ..$ content        : chr [1:183] "#DIV/0!" "57" "1" "55" ...
#>   ..$ formula        : chr [1:183] "1/0" NA NA NA ...
#>   ..$ formula_type   : chr [1:183] NA NA NA NA ...
#>   ..$ formula_ref    : chr [1:183] NA NA NA NA ...
#>   ..$ formula_group  : int [1:183] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ type           : chr [1:183] "e" "s" "b" "s" ...
#>   ..$ data_type      : chr [1:183] "error" "character" "logical" "character" ...
#>   ..$ error          : chr [1:183] "#DIV/0!" NA NA NA ...
#>   ..$ logical        : logi [1:183] NA NA TRUE NA FALSE NA ...
#>   ..$ numeric        : num [1:183] NA NA NA NA NA ...
#>   ..$ date           : POSIXct[1:183], format: NA NA ...
#>   ..$ character      : chr [1:183] NA "Error" NA "Boolean (true)" ...
#>   ..$ comment        : chr [1:183] NA NA NA NA ...
#>   ..$ height         : num [1:183] 14.2 14.2 14.2 14.2 14.2 ...
#>   ..$ width          : num [1:183] 19.4 68 19.4 68 19.4 ...
#>   ..$ style_format   : chr [1:183] "Normal" "Normal" "Normal" "Normal" ...
#>   ..$ local_format_id: int [1:183] 1 1 1 1 1 1 1 1 5 1 ...
#>  $ 1~`!@#$%^&()_-+={}|;"'<,>.£¹÷¦°:Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	2 obs. of  20 variables:
#>   ..$ address        : chr [1:2] "B2" "B6"
#>   ..$ row            : int [1:2] 2 6
#>   ..$ col            : int [1:2] 2 2
#>   ..$ content        : chr [1:2] "30" "31"
#>   ..$ formula        : chr [1:2] NA NA
#>   ..$ formula_type   : chr [1:2] NA NA
#>   ..$ formula_ref    : chr [1:2] NA NA
#>   ..$ formula_group  : int [1:2] NA NA
#>   ..$ type           : chr [1:2] "s" "s"
#>   ..$ data_type      : chr [1:2] "character" "character"
#>   ..$ error          : chr [1:2] NA NA
#>   ..$ logical        : logi [1:2] NA NA
#>   ..$ numeric        : num [1:2] NA NA
#>   ..$ date           : POSIXct[1:2], format: NA NA
#>   ..$ character      : chr [1:2] "empty first row and col" "This worksheet has a nasty name but can be imported nevertheless"
#>   ..$ comment        : chr [1:2] "nacnudus:" NA
#>   ..$ height         : num [1:2] 14.2 14.2
#>   ..$ width          : num [1:2] 8.38 8.38
#>   ..$ style_format   : chr [1:2] "Normal" "Normal"
#>   ..$ local_format_id: int [1:2] 1 1

# Specific sheet either by position or by name
str(tidy_xlsx(examples, 2)$data)

#> List of 1
#>  $ 1~`!@#$%^&()_-+={}|;"'<,>.£¹÷¦°:Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	2 obs. of  20 variables:
#>   ..$ address        : chr [1:2] "B2" "B6"
#>   ..$ row            : int [1:2] 2 6
#>   ..$ col            : int [1:2] 2 2
#>   ..$ content        : chr [1:2] "30" "31"
#>   ..$ formula        : chr [1:2] NA NA
#>   ..$ formula_type   : chr [1:2] NA NA
#>   ..$ formula_ref    : chr [1:2] NA NA
#>   ..$ formula_group  : int [1:2] NA NA
#>   ..$ type           : chr [1:2] "s" "s"
#>   ..$ data_type      : chr [1:2] "character" "character"
#>   ..$ error          : chr [1:2] NA NA
#>   ..$ logical        : logi [1:2] NA NA
#>   ..$ numeric        : num [1:2] NA NA
#>   ..$ date           : POSIXct[1:2], format: NA NA
#>   ..$ character      : chr [1:2] "empty first row and col" "This worksheet has a nasty name but can be imported nevertheless"
#>   ..$ comment        : chr [1:2] "nacnudus:" NA
#>   ..$ height         : num [1:2] 14.2 14.2
#>   ..$ width          : num [1:2] 8.38 8.38
#>   ..$ style_format   : chr [1:2] "Normal" "Normal"
#>   ..$ local_format_id: int [1:2] 1 1

str(tidy_xlsx(examples, "Sheet1")$data)

#> List of 1
#>  $ Sheet1:Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	183 obs. of  20 variables:
#>   ..$ address        : chr [1:183] "A1" "B1" "A2" "B2" ...
#>   ..$ row            : int [1:183] 1 1 2 2 3 3 4 4 5 5 ...
#>   ..$ col            : int [1:183] 1 2 1 2 1 2 1 2 1 2 ...
#>   ..$ content        : chr [1:183] "#DIV/0!" "57" "1" "55" ...
#>   ..$ formula        : chr [1:183] "1/0" NA NA NA ...
#>   ..$ formula_type   : chr [1:183] NA NA NA NA ...
#>   ..$ formula_ref    : chr [1:183] NA NA NA NA ...
#>   ..$ formula_group  : int [1:183] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ type           : chr [1:183] "e" "s" "b" "s" ...
#>   ..$ data_type      : chr [1:183] "error" "character" "logical" "character" ...
#>   ..$ error          : chr [1:183] "#DIV/0!" NA NA NA ...
#>   ..$ logical        : logi [1:183] NA NA TRUE NA FALSE NA ...
#>   ..$ numeric        : num [1:183] NA NA NA NA NA ...
#>   ..$ date           : POSIXct[1:183], format: NA NA ...
#>   ..$ character      : chr [1:183] NA "Error" NA "Boolean (true)" ...
#>   ..$ comment        : chr [1:183] NA NA NA NA ...
#>   ..$ height         : num [1:183] 14.2 14.2 14.2 14.2 14.2 ...
#>   ..$ width          : num [1:183] 19.4 68 19.4 68 19.4 ...
#>   ..$ style_format   : chr [1:183] "Normal" "Normal" "Normal" "Normal" ...
#>   ..$ local_format_id: int [1:183] 1 1 1 1 1 1 1 1 5 1 ...

# Data (cell values)
x <- tidy_xlsx(examples)
str(x$data$Sheet1)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	183 obs. of  20 variables:
#>  $ address        : chr  "A1" "B1" "A2" "B2" ...
#>  $ row            : int  1 1 2 2 3 3 4 4 5 5 ...
#>  $ col            : int  1 2 1 2 1 2 1 2 1 2 ...
#>  $ content        : chr  "#DIV/0!" "57" "1" "55" ...
#>  $ formula        : chr  "1/0" NA NA NA ...
#>  $ formula_type   : chr  NA NA NA NA ...
#>  $ formula_ref    : chr  NA NA NA NA ...
#>  $ formula_group  : int  NA NA NA NA NA NA NA NA NA NA ...
#>  $ type           : chr  "e" "s" "b" "s" ...
#>  $ data_type      : chr  "error" "character" "logical" "character" ...
#>  $ error          : chr  "#DIV/0!" NA NA NA ...
#>  $ logical        : logi  NA NA TRUE NA FALSE NA ...
#>  $ numeric        : num  NA NA NA NA NA ...
#>  $ date           : POSIXct, format: NA NA ...
#>  $ character      : chr  NA "Error" NA "Boolean (true)" ...
#>  $ comment        : chr  NA NA NA NA ...
#>  $ height         : num  14.2 14.2 14.2 14.2 14.2 ...
#>  $ width          : num  19.4 68 19.4 68 19.4 ...
#>  $ style_format   : chr  "Normal" "Normal" "Normal" "Normal" ...
#>  $ local_format_id: int  1 1 1 1 1 1 1 1 5 1 ...

# Formatting
str(x$formats$local)

#> List of 4
#>  $ numFmt: chr [1:76] "General" "d-mmm-yy" "[$-1409]d\\ mmmm\\ yyyy;@" "yyyy\\ mmmm\\ dddd" ...
#>  $ font  :List of 10
#>   ..$ bold     : logi [1:76] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ italic   : logi [1:76] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ underline: chr [1:76] NA NA NA NA ...
#>   ..$ strike   : logi [1:76] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ vertAlign: chr [1:76] NA NA NA NA ...
#>   ..$ size     : int [1:76] 11 11 11 11 11 11 11 11 11 11 ...
#>   ..$ color    :List of 4
#>   .. ..$ rgb    : chr [1:76] NA NA NA NA ...
#>   .. ..$ theme  : int [1:76] NA NA NA NA NA NA NA NA 5 NA ...
#>   .. ..$ indexed: int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. ..$ tint   : num [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ name     : chr [1:76] "Arial" "Arial" "Arial" "Arial" ...
#>   ..$ family   : int [1:76] 2 2 2 2 2 2 2 2 2 2 ...
#>   ..$ scheme   : chr [1:76] NA NA NA NA ...
#>  $ fill  :List of 2
#>   ..$ patternFill :List of 3
#>   .. ..$ fgColor    :List of 4
#>   .. .. ..$ rgb    : chr [1:76] NA NA NA NA ...
#>   .. .. ..$ theme  : int [1:76] NA NA NA NA NA NA NA 5 NA NA ...
#>   .. .. ..$ indexed: int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ tint   : num [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. ..$ bgColor    :List of 4
#>   .. .. ..$ rgb    : chr [1:76] NA NA NA NA ...
#>   .. .. ..$ theme  : int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ indexed: int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ tint   : num [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. ..$ patternType: chr [1:76] "none" "none" "none" "none" ...
#>   ..$ gradientFill:List of 8
#>   .. ..$ type   : chr [1:76] NA NA NA NA ...
#>   .. ..$ degree_: int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. ..$ left   : num [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. ..$ right  : num [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. ..$ top    : num [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. ..$ bottom : num [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. ..$ color1 :List of 4
#>   .. .. ..$ rgb    : chr [1:76] NA NA NA NA ...
#>   .. .. ..$ theme  : int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ indexed: int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ tint   : num [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. ..$ color2 :List of 4
#>   .. .. ..$ rgb    : chr [1:76] NA NA NA NA ...
#>   .. .. ..$ theme  : int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ indexed: int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ tint   : num [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>  $ border:List of 12
#>   ..$ diagonalDown: int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ diagonalUp  : int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ outline     : int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ left        :List of 2
#>   .. ..$ style: chr [1:76] NA NA NA NA ...
#>   .. ..$ color:List of 4
#>   .. .. ..$ rgb    : chr [1:76] NA NA NA NA ...
#>   .. .. ..$ theme  : int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ indexed: int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ tint   : num [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ right       :List of 2
#>   .. ..$ style: chr [1:76] NA NA NA NA ...
#>   .. ..$ color:List of 4
#>   .. .. ..$ rgb    : chr [1:76] NA NA NA NA ...
#>   .. .. ..$ theme  : int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ indexed: int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ tint   : num [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ start       :List of 2
#>   .. ..$ style: chr [1:76] NA NA NA NA ...
#>   .. ..$ color:List of 4
#>   .. .. ..$ rgb    : chr [1:76] NA NA NA NA ...
#>   .. .. ..$ theme  : int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ indexed: int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ tint   : num [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ end         :List of 2
#>   .. ..$ style: chr [1:76] NA NA NA NA ...
#>   .. ..$ color:List of 4
#>   .. .. ..$ rgb    : chr [1:76] NA NA NA NA ...
#>   .. .. ..$ theme  : int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ indexed: int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ tint   : num [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ top         :List of 2
#>   .. ..$ style: chr [1:76] NA NA NA NA ...
#>   .. ..$ color:List of 4
#>   .. .. ..$ rgb    : chr [1:76] NA NA NA NA ...
#>   .. .. ..$ theme  : int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ indexed: int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ tint   : num [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ bottom      :List of 2
#>   .. ..$ style: chr [1:76] NA NA NA NA ...
#>   .. ..$ color:List of 4
#>   .. .. ..$ rgb    : chr [1:76] NA NA NA NA ...
#>   .. .. ..$ theme  : int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ indexed: int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ tint   : num [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ diagonal    :List of 2
#>   .. ..$ style: chr [1:76] NA NA NA NA ...
#>   .. ..$ color:List of 4
#>   .. .. ..$ rgb    : chr [1:76] NA NA NA NA ...
#>   .. .. ..$ theme  : int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ indexed: int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ tint   : num [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ vertical    :List of 2
#>   .. ..$ style: chr [1:76] NA NA NA NA ...
#>   .. ..$ color:List of 4
#>   .. .. ..$ rgb    : chr [1:76] NA NA NA NA ...
#>   .. .. ..$ theme  : int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ indexed: int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ tint   : num [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ horizontal  :List of 2
#>   .. ..$ style: chr [1:76] NA NA NA NA ...
#>   .. ..$ color:List of 4
#>   .. .. ..$ rgb    : chr [1:76] NA NA NA NA ...
#>   .. .. ..$ theme  : int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ indexed: int [1:76] NA NA NA NA NA NA NA NA NA NA ...
#>   .. .. ..$ tint   : num [1:76] NA NA NA NA NA NA NA NA NA NA ...

# The formats of particular cells can be retrieved like this:

Sheet1 <- x$data$Sheet1
x$formats$style$font$bold[Sheet1$style_format]

#>     Normal     Normal     Normal     Normal     Normal     Normal     Normal 
#>      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE 
#>     Normal     Normal     Normal     Normal     Normal     Normal     Normal 
#>      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE 
#>     Normal     Normal     Normal     Normal     Normal     Normal     Normal 
#>      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE 
#>     Normal    Percent     Normal     Normal     Normal     Normal     Normal 
#>      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE 
#>     Normal     Normal     Normal     Normal     Normal     Normal     Normal 
#>      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE 
#>     Normal     Normal     Normal     Normal     Normal     Normal     Normal 
#>      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE 
#>     Normal     Normal     Normal     Normal     Normal     Normal     Normal 
#>      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE 
#>     Normal     Normal     Normal     Normal     Normal     Normal     Normal 
#>      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE 
#>     Normal     Normal     Normal     Normal        Bad     Normal        Bad 
#>      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE 
#>     Normal     Normal     Normal     Normal     Normal     Normal     Normal 
#>      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE 
#>     Normal     Normal stylealign     Normal     Normal     Normal     Normal 
#>      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE 
#>     Normal     Normal     Normal     Normal     Normal     Normal     Normal 
#>      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE 
#>     Normal     Normal     Normal     Normal     Normal     Normal     Normal 
#>      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE 
#>     Normal     Normal     Normal     Normal     Normal     Normal     Normal 
#>      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE 
#>     Normal   Normal 2     Normal   Normal 2   Normal 2   Normal 2   Normal 2 
#>      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE 
#>   Normal 2   Normal 2   Normal 2   Normal 2     Normal   Normal 2     Normal 
#>      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE 
#>   Normal 2     Normal   Normal 2     Normal   Normal 2     Normal   Normal 2 
#>      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE 
#>     Normal   Normal 2     Normal   Normal 2     Normal   Normal 2     Normal 
#>      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE 
#>   Normal 2  Hyperlink   Normal 2  Hyperlink   Normal 2     Normal   Normal 2 
#>      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE 
#>     Normal     Normal     Normal     Normal     Normal     Normal     Normal 
#>      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE 
#>     Normal     Normal     Normal     Normal     Normal     Normal     Normal 
#>      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE 
#>     Normal     Normal     Normal     Normal     Normal     Normal     Normal 
#>      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE 
#>     Normal     Normal     Normal     Normal     Normal     Normal     Normal 
#>      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE 
#>     Normal     Normal     Normal     Normal     Normal     Normal     Normal 
#>      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE 
#>     Normal     Normal     Normal     Normal     Normal     Normal     Normal 
#>      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE 
#>     Normal     Normal     Normal     Normal     Normal     Normal     Normal 
#>      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE      FALSE 
#>     Normal 
#>      FALSE 

x$formats$local$font$bold[Sheet1$local_format_id]

#>   [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [13] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [25] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [37] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [49] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE
#>  [61] FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [73] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [85] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE
#>  [97] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [109] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [121] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [133] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [145] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [157] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [169] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [181] FALSE FALSE FALSE

# To filter for cells of a particular format, first filter the formats to get
# the relevant indices, and then filter the cells by those indices.
bold_indices <- which(x$formats$local$font$bold)
Sheet1[Sheet1$local_format_id %in% bold_indices, ]

#> # A tibble: 3 × 20
#>   address   row   col content formula formula_type formula_ref formula_group
#>     <chr> <int> <int>   <chr>   <chr>        <chr>       <chr>         <int>
#> 1     A28    28     1       0    <NA>         <NA>        <NA>            NA
#> 2     A32    32     1       6    <NA>         <NA>        <NA>            NA
#> 3     A48    48     1      13    <NA>         <NA>        <NA>            NA
#> # ... with 12 more variables: type <chr>, data_type <chr>, error <chr>,
#> #   logical <lgl>, numeric <dbl>, date <dttm>, character <chr>, comment <chr>,
#> #   height <dbl>, width <dbl>, style_format <chr>, local_format_id <int>