List sheets in an xlsx (Excel) file

xlsx_sheets() returns the names of the sheets in a workbook, as a character vector. They are in the same order as they appear in the spreadsheet when it is opened with a spreadsheet application like Excel or LibreOffice.

Usage

xlsx_sheet_names(path, check_filetype = TRUE)

Arguments

path

Path to the xlsx file.

check_filetype

Logical. Whether to check that the filetype is xlsx (or xlsm) by looking at the file itself, rather than using the filename extension.

Value

A character vector of the names of the worksheets in the file.

Examples

examples <- system.file("extdata/examples.xlsx", package = "tidyxl")
xlsx_sheet_names(examples)
#> [1] "Sheet1"                      "1~`!@#$%^&()_-+={}|;\"'<,>."
#> [3] "E09904.2"                    "gradientFill"