Import data validation rules of cells in xlsx (Excel) files

xlsx_validation() returns the data validation rules applied to cells in xlsx (Excel) files. Data validation rules control what constants can be entered into a cell, e.g. any whole number between 0 and 9, or one of several values from another part of the spreadsheet.

Each data validation rule can only apply to cells on one worksheet, so xlsx_validation() returns a named list of data frames, one for each worksheet, describing the rules on that worksheet.

xlsx_validation(path, sheets = NA)

Arguments

path
Path to the xlsx file.

sheets
Sheets to read. Either a character vector (the names of the sheets), an integer vector (the positions of the sheets), or NA (default, all sheets).

Value

A named list of data frames, one for each worksheet in the file. Each data frame has the following columns.

ref Comma-delimited cell addresses to which the rules apply, e.g. A106 or A115,A121:A122`.

type Data type of input, one ofwhole,decimal,list,date,time,textLength,custom, andwhole`.

operator Unless type is list or custom, then operator is one of between, notBetween, equal, notEqual, greaterThan, lessThan, greaterThanOrEqual, lessthanOrEqual.

formula1 If type is list, then a range of cells whose values are allowed by the rule. If type is custom, then a formula to determine alowable values. Otherwise, a cell address or constant, coerced to character. Dates and times are formatted like "2017-01-27 13:30:45". Times without dates are formatted like "13:30:45".

formula2 If operator is between or notBetween, then a cell address or constant as with formula1, otherwise NA.

allow_blank Boolean, whether or not the rule allows blanks.

show_input_message Boolean, whether or not the rule shows a message when the user begins entering a value.

prompt_title Text to appear in the title bar of a popup message box when the user begins entering a value.

prompt_body Text to appear in a popup message box when the user begins entering a value. When NA, then some default text is shown.

show_error_message Boolean, whether or not the rule shows a message when the user has entered a forbidden value. When NA, then some default text is shown.

error_title Text to appear in the title bar of a popup message box when the user enters a forbidden value. When NA, then some default text is shown.

error_body Text to appear in a popup message box when the user enters a forbidden value. When NA, then some default text is shown.

error_symbol Name of a symbol to appear in the popup error message when the user enters a forbidden value.

Examples

examples <- system.file("extdata/examples.xlsx", package = "tidyxl")
xlsx_validation(examples)

#> $Sheet1
#> # A tibble: 15 x 13
#>               ref       type           operator            formula1
#>             <chr>      <chr>              <chr>               <chr>
#>  1           A106      whole            between                   0
#>  2           A107    decimal         notBetween                   0
#>  3           A108       list               <NA>              $B$108
#>  4           A109       list               <NA>              $B$108
#>  5           A110       date            between 2017-01-01 00:00:00
#>  6           A111       time            between            00:00:00
#>  7           A112 textLength            between                   0
#>  8           A113     custom               <NA>     A113<=LEN(B113)
#>  9           A114      whole         notBetween                   0
#> 10 A115,A121:A122      whole              equal                   0
#> 11           A116      whole           notEqual                   0
#> 12           A117      whole        greaterThan                   0
#> 13           A118      whole           lessThan                   0
#> 14           A119      whole greaterThanOrEqual                   0
#> 15           A120      whole    lessThanOrEqual                   0
#> # ... with 9 more variables: formula2 <chr>, allow_blank <lgl>,
#> #   show_input_message <lgl>, prompt_title <chr>, prompt_body <chr>,
#> #   show_error_message <lgl>, error_title <chr>, error_body <chr>,
#> #   error_symbol <chr>
#> 
#> $`1~`!@#$%^&()_-+={}|;"'<,>.£¹÷¦°`
#> # A tibble: 0 x 13
#> # ... with 13 variables: ref <chr>, type <chr>, operator <chr>, formula1 <chr>,
#> #   formula2 <chr>, allow_blank <lgl>, show_input_message <lgl>,
#> #   prompt_title <chr>, prompt_body <chr>, show_error_message <lgl>,
#> #   error_title <chr>, error_body <chr>, error_symbol <chr>
#> 
#> $E09904.2
#> # A tibble: 0 x 13
#> # ... with 13 variables: ref <chr>, type <chr>, operator <chr>, formula1 <chr>,
#> #   formula2 <chr>, allow_blank <lgl>, show_input_message <lgl>,
#> #   prompt_title <chr>, prompt_body <chr>, show_error_message <lgl>,
#> #   error_title <chr>, error_body <chr>, error_symbol <chr>
#> 

xlsx_validation(examples, 1)[["Sheet1"]]

#> # A tibble: 15 x 13
#>               ref       type           operator            formula1
#>             <chr>      <chr>              <chr>               <chr>
#>  1           A106      whole            between                   0
#>  2           A107    decimal         notBetween                   0
#>  3           A108       list               <NA>              $B$108
#>  4           A109       list               <NA>              $B$108
#>  5           A110       date            between 2017-01-01 00:00:00
#>  6           A111       time            between            00:00:00
#>  7           A112 textLength            between                   0
#>  8           A113     custom               <NA>     A113<=LEN(B113)
#>  9           A114      whole         notBetween                   0
#> 10 A115,A121:A122      whole              equal                   0
#> 11           A116      whole           notEqual                   0
#> 12           A117      whole        greaterThan                   0
#> 13           A118      whole           lessThan                   0
#> 14           A119      whole greaterThanOrEqual                   0
#> 15           A120      whole    lessThanOrEqual                   0
#> # ... with 9 more variables: formula2 <chr>, allow_blank <lgl>,
#> #   show_input_message <lgl>, prompt_title <chr>, prompt_body <chr>,
#> #   show_error_message <lgl>, error_title <chr>, error_body <chr>,
#> #   error_symbol <chr>

xlsx_validation(examples, "Sheet1")[[1]]

#> # A tibble: 15 x 13
#>               ref       type           operator            formula1
#>             <chr>      <chr>              <chr>               <chr>
#>  1           A106      whole            between                   0
#>  2           A107    decimal         notBetween                   0
#>  3           A108       list               <NA>              $B$108
#>  4           A109       list               <NA>              $B$108
#>  5           A110       date            between 2017-01-01 00:00:00
#>  6           A111       time            between            00:00:00
#>  7           A112 textLength            between                   0
#>  8           A113     custom               <NA>     A113<=LEN(B113)
#>  9           A114      whole         notBetween                   0
#> 10 A115,A121:A122      whole              equal                   0
#> 11           A116      whole           notEqual                   0
#> 12           A117      whole        greaterThan                   0
#> 13           A118      whole           lessThan                   0
#> 14           A119      whole greaterThanOrEqual                   0
#> 15           A120      whole    lessThanOrEqual                   0
#> # ... with 9 more variables: formula2 <chr>, allow_blank <lgl>,
#> #   show_input_message <lgl>, prompt_title <chr>, prompt_body <chr>,
#> #   show_error_message <lgl>, error_title <chr>, error_body <chr>,
#> #   error_symbol <chr>